Points Of Interest sorted in descending rating order; moved Points Of Interest into LocationProfile folder
