+++ src/components/LocationProfile/PointsOfInterest.jsx
@@ -1,10 +1,11 @@
 import React from "react";
 
 const PointsOfInterest = (props) => {
+  var sorted = props.pointsOfInterest.sort((a, b) => {return b.rating - a.rating});
   return (
     <div className="points-of-interest-section">
       <h2>Top Places To Visit In {props.location}:</h2>
-      {props.pointsOfInterest.map((place, ind) => (
+      {sorted.map((place, ind) => (
         <div className="point-of-interest" key={ind}>
           <h3>{place.name}</h3>
           <p>Address: {place.formatted_address}</p>
